fix pagination bug and make filters dynamic
