+++ app/Http/Controllers/MarketplaceController.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Http\Controllers;
+
+class MarketplaceController extends Controller
+{
+    public function index()
+    {
+        return view('marketplace.index');
+    }
+}
