+++ resources/views/livewire/components/market/show-listing-modal-component.blade.php
